refactor: making match page more generic & re-usable
