+++ src/components/RecordingSection/RecordingSectionRecording.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { useZustandStore } from "@/lib/useZustandStore";
+import RecordingSectionGeneric from "@/components/RecordingSection/RecordingSectionGeneric";
+
+export default function RecordingSectionRecording() {
+  const recordingUlid = useZustandStore((state) => state.recordingUlid);
+  const autoSaveFile = useZustandStore((state) => state.autoSaveFile);
+
+  return (
+    <RecordingSectionGeneric
+      recordingUlid={recordingUlid}
+      autoSaveFile={autoSaveFile}
+    />
+  );
+}
